LPS-151267 search-experiences-service:move/rename class files
https://issues.liferay.com/browse/LPS-151267
