+++ modules/dxp/apps/search-experiences/search-experiences-service/src/main/java/com/liferay/search/experiences/internal/upgrade/SXPUpgradeStepRegistrator.java
@@ -16,9 +16,7 @@ package com.liferay.search.experiences.internal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.BaseExternalReferenceCodeUpgradeProcess;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.search.experiences.internal.upgrade.v1_1_0.SXPBlueprintKeyVersionUpgradeProcess;
 import com.liferay.search.experiences.internal.upgrade.v1_1_0.SXPBlueprintUpgradeProcess;
-import com.liferay.search.experiences.internal.upgrade.v1_1_0.SXPElementKeyVersionUpgradeProcess;
 import com.liferay.search.experiences.internal.upgrade.v1_1_0.SXPElementUpgradeProcess;
 
 import org.osgi.service.component.annotations.Component;
@@ -41,7 +39,7 @@ public class SXPUpgradeStepRegistrator implements UpgradeStepRegistrator {
 		registry.register(
 			"1.1.0", "1.2.0",
 			new BaseExternalReferenceCodeUpgradeProcess() {
-
+				
 				@Override
 				protected String[][] getTableAndPrimaryKeyColumnNames() {
 					return new String[][] {
